Clean Makefile.bundles copy method
Formerly, Makefile.bundles copied files as an action
of the "all" phony target, causing files to be copied
even if nothing has to be done. Files to copy are now
regular targets of the "all" target, with a proper
rule to handle the copy.

menhir bundle is now considered only if it is activated
by configure.

These changes enable make to say "Nothing to be done"
if everything is already built.