allow EndNode matching
The control-flow graph for a #define with a statement uses EndNode
rather than Exit as the endpoint.  Previously there was no predicate to
match an EndNode.  This makes the Exit predicate also match an EndNode
and adds EndNode to the set of nodes that can match a MetaRuleElem (like
after, fallthrough, etc).  It seems that matching a statement in the if
body such as if (x) return; else x = 12; triggers an inconsistent path
warning.  So there is probably something more to fix up.