special case for adding whitespace
Adding strings seems to cause the adjacent added code to not be
considered to be able to be eg a complex statement, and instead it ends
up as a sequence of tokens.  This is due to the function
contains_added_string in the file context_neg.ml.  At least in the case
where the added string just contains whitespace, this does not seem
necessary.