disallow binding ifdef and cpp lines to field metavariables
Ifdef and cpp lines are filtered out of the counting of field lists.
This introduces an ambiguity if the nth field ends in an ifdef.  Should
a subsequent field metavariable match the next real field or the endif.
Currently it was matching both.  But it is not intuitive for MetaFields
to get bound to endif.  So just forbid this.