enable adding includes in header files
Header files often contain the idiom

to avoid double inclusion.  Normally, parsing_hacks avoids adding an
include after an include matching an anonymous patttern (ie #include
<...>) that is inside an ifdef, because it would be better to have the
added include at top level.  But in a header with the above pattern
there is no code and thus no include at top level.  So we just ignore
the above pattern in calculating whether an include is in an ifdef.