better sorting of fake tokens
It seems that fake tokens have an offset that should be -1 if the fake
token is before a real token and the width of the real token if the fake
token is after the real one.  Use this to compare fake and real tokens.
It seems that having the fake token that comes before a declaration
actually come before the declaration is very important to the management
of multidecls, as the danger start should be on the fake token, and nto
the first token on the type.