correct fake positions added after conditionals and loops
Fake tokens are supposed to have a position, which is the position of
the token that motivates their existence, and an offset that allows the
token to be ordered with the rest.  Update the positions for the fake
tokens at the end of a conditional or loop so that the position is not 0
and the offset -1.  Instead use the largest position of the preceeding
tokens.