add EOL for #pragma
This is needed to allow dropping pragmas that appear in the middle of
code and interfere with parsing.  See parse_c.ml.

Also, make the last ii of PrePragma available to define_line_2 so that
the EOL gets a character position just past the end of the complete
pragma line.