adjust choice of ifdef branch
Commenting out ifdefs is only useful if the code in the branches is
sequencible, ie if one can b placed after the other.  If not, it is
better to just take the first branch.  Increase the set of cases in
which this is done, by looking for sequencible indicators at the end of
the last branch and for cpp directives in the code (eg #define can end
in anything).