adjust undanger_untouched
Undanger_untouched breaks a line apart at commas, to find each declared
variable, and check that there is some context code in each case,
meaning that at least part of the variable is still there.  This should
skip over the Danger parts at the beginning of the declaration, ie the
type information, as that has nothing to do with whether variables are
still there.