Avoid duplication on changes on types in the presence of multidecls.
A change on a type has to fold over the contents of each CFG node to
find mentions of the type.  This is done by cocciTy.  In that function,
if we encounter a declaration, we only want to continue folding into the
first type, as all of the types in a given declaration are shared. The
visitor for C code then has to be extended to be able to visit a
onedecl, but ignore the type.