address missing newlines around commented out code in + metavariables
The problem is with the align.c test.  The ALIGN(32) is commented out,
ubt it is getting matched inside a field list metavariable and we would
like to keep it around in the added code.  This is done by allowing
al_comments to keep such commented out code.  That will impact how
metavariables are matched, but perhaps such code really is part of the
metavariables.  But if we keep the code, we may need to keep the
newlines around it.

The problem is that commented code is dealt with by each token just
taking all of the subsequent commented out code.  A better solution
would be to take all the code up to the final newline, if any, and then
to also take the preceding commented out code up to the last newline.
But this may have unforeseen consequences, so in the short term we take
this less intrusive solution.