better manage initialisation to affectation
The complete node should not be replaced because it might also be being
matched as an initialization, and this will result in a bad match in the
transformation process.  It should be ok to use the original node,
because the tokens that are modified are inside of this.

Problem reported by Markus Elfring.