GADGET3 AoS->SoA semantic patches as tests
These files have been developed to assist a "semantic patch"-driven
restructuring of the data layout of the GADGET simulation code:
from an Arrays-of-Structures layout to a Structures-of-Arrays one.

In terms of C code, it means changing the layout from e.g.:

 // arrays of structs (here P):
 struct particle_data { MyFloat Mass; short int Type; }  *P;
 // ...
 for(i=0;i<NumPart;i++){ in->Mass = P[i].Mass; in->Type = P[i].Type; }

into:
 // structures of arrays (P_Mass, P_type):
 MyFloat*   P_Mass; short int* P_Type;
 // ...
 for(i=0;i<NumPart;i++) { in->Mass = P_Mass[i]; in->Type = P_Type[i]; }

These files can serve here as Coccinelle tests as well.

Thanks to Antonio Ragagnin for sharing GADGET-specific knowledge, and
Julia Lawall for helping Coccinelle-side.

Signed-off-by: Michele Martone <michelemartone@users.sourceforge.net>