extract real comments
The comment extraction functions provided by ast_c return all tokens
between the code elements and not just the ones that are actual
comments.  The non comment tokens were being filtered out in the
conversion to script variables, but only for the string representations,
not for the internal representations.  Change the metavariable binding
so that it only contains information about real comments.  Note that the
code making the script bindings does not change, because it is still
necessary to extract the strings from their token data structures.