fix dependency for parsing_cocci/cocci_args
parsing_cocci/cocci_args depends on the Stdcompat library, but was not
part of the STDCOMPAT_USERS variable. This results in build failures if
Make is run using -j to build in parallel.

Fix this by ensuring that parsing_cocci/cocci_args is included in the
STDCOMPAT_USERS variable, ensuring that a proper dependency on
STDCOMPAT_LIB is created by Make.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>