parsing_cocci: parse_cocci: Fix struct pointer funcall regression
This is a fix in response to the following bug report:

https://www.mail-archive.com/cocci@systeme.lip6.fr/msg07332.html

The following commit:

	c280375635f62dfbe052709e4e47a82140d32ce5

Introduces a regression in the following SmPL use case:

	@@
	struct s *x;
	@@

	x->func();

Where x is a pointer to a struct and func is a function belonging to
said struct. The faulty commit mislabels func as a function prototype
due to a missing case in the function-prototype-detection match ladder.

Add a case in the match ladder for a struct pointer function call.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>