cocci: Add parameter attributes to C and SmPL AST
This is a combination of 28 commits.
This is the 1st commit message:

cocci: Add parameter attributes to C AST

With extended support of attributes, parameter attributes are needed in
the C AST so that the nocast test case does not break.

Add the p_attr field to parameter in the C AST.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #2:

parsing_c: parser: Place parameter attributes in C AST

Parameter attributes are needed so as to not break the nocast test case
when attributes are fully supported. Add parameter attributes to the C
AST.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #3:

parsing_c: visitor_c: Visit parameter attributes

As attributes are added to the parameter type, have the C AST visitor
visit these attributes as well.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #4:

parsing_c: unparse_hrule: Add parameter attributes in record

Parameter attributes are added to the C AST. Initialize the parameter
attributes field as empty in a case in unparse_hrule.ml.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #5:

parsing_c: type_c: Add parameter attributes to record

Paramter attributes are added to the C AST. Reflect this change in a
case in type_c.ml.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #6:

engine: cocci_vs_c: "Match" parameter attributes

Parameter attributes are added to the C AST, but not to the SmPL AST.
Once parameter attributes are added to SmPL as well, they can actually
be matched, but for now a call to attribute_list is required to
correctly remove attributes. Attributes from a parameter are removed
when the whole parameter is removed, replaced or otherwise.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #7:

parsing_cocci: ast0_cocci: Add parameter attributes

Add parameter attributes to AST0 of SmPL. This is a list of attributes
in the VoidParam and Param types of AST0.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #8:

parsing_cocci: parser: Parse Parameter attributes

Parameter attributes are added to the SmPL AST. Parse these attributes
correctly in the SmPL parser. The added production only supports
attributes after the type or the type and identifier.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #9:

parsing_cocci: visitor_ast0: Visit Parameter attributes

Parameter attributes are added to AST0 of SmPL. Visit these attributes
in the AST0 visitor.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #10:

parsing_cocci: arity: Reflect Parameter attributes

Parameter attributes are added to the SmPL AST. Reflect these changes in
arity.ml.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #11:

parsing_cocci: check_meta: Reflect Parameter attributes

Parameter attributes are added to the SmPL AST. Reflect these changes in
check_meta.ml.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #12:

parsing_cocci: compute_lines: Reflect Parameter attributes

Parameter attributes are added to the SmPL AST. Reflect these changes in
compute_lines.ml.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #13:

parsing_cocci: context_neg: Reflect Parameter attributes

Parameter attributes are added to the SmPL AST. Reflect these changes in
context_neg.ml.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #14:

parsing_cocci: function_prototypes: Reflect Parameter attributes

Parameter attributes are added to the SmPL AST. Reflect these changes in
function_prototypes.ml.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #15:

parsing_cocci: index: Reflect Parameter attributes

Parameter attributes are added to the SmPL AST. Reflect these changes in
index.ml.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #16:

parsing_cocci: iso_pattern: Reflect Parameter attributes

Parameter attributes are added to the SmPL AST. Reflect these changes in
iso_pattern.ml.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #17:

parsing_cocci: type_infer: Reflect Parameter attributes

Parameter attributes are added to the SmPL AST. Reflect these changes in
type_infer.ml.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #18:

parsing_cocci: unparse_ast0: Reflect Parameter attributes

Parameter attributes are added to the SmPL AST. Reflect these changes in
unparse_ast0.ml.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #19:

parsing_c: unparse_cocci: Reflect Parameter attributes

Parameter attributes are added to the SmPL AST. Print these attributes
correctly in unparse_cocci.ml.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #20:

parsing_cocci: ast_cocci: Add Parameter attributes

Add parameter attributes to the SmPL AST. This is a list of attributes
in the VoidParam and Param types of the SmPL AST.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #21:

parsing_cocci: visitor_ast: Visit Parameter attributes

Parameter attributes are added to the SmPL AST. Visit these attributes in
the AST visitor.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #22:

parsing_cocci: ast0toast: Reflect Parameter attributes

Parameter attributes are added to the SmPL AST. Reflect these changes in
ast0toast.ml.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #23:

parsing_cocci: disjdistr: Reflect Parameter attributes

Parameter attributes are added to the SmPL AST. Reflect these changes in
disjdistr.ml.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #24:

parsing_cocci: pretty_print_cocci: Reflect Parameter attributes

Parameter attributes are added to the SmPL AST. Print these attributes
correctly in pretty_print_cocci.ml.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #25:

parsing_cocci: unify_ast: Reflect Parameter attributes

Parameter attributes are added to the SmPL AST. Reflect these changes in
unify_ast.ml.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #26:

engine: cocci_vs_c: Match Parameter attributes

Parameter attributes are added to the C and SmPL AST. Match the
attributes correctly in cocci_vs_c.ml.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #27:

ocaml: coccilib: Reflect Parameter attributes

Parameter attributes are added to the SmPL AST. Reflect these changes in
coccilib.mli.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>

This is the commit message #28:

engine: c_vs_c: Match parameter attributes

Parameter attributes are added to the C AST. Add an expression to check
if attributes are equal on both sides to the parameter case in c_vs_c.

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>