more informative error message on malformed cfg
Include some information about the affected code in the error message.

The improved error message can be seen by running the following in
the tests directory

spatch efi3.cocci efi3.c

This causes an error because of the exit: label in the statement
expression. Labels are made in advance in control_flow_build and then
here checked for their reachability.