improve andany optimization
The idea behind AndAny is that when we have eg { <... P ...> } we just
want to find all of the matches of P within the braces.  We don't care
about following all control-flow paths.  Ignoring the specific
control-flow paths can be a large performance improvement, especially if
loops or a long sequence of ifs are involved.

The previous solution was to find all of the places where P matched and
conjunct them individually one by one with the starting point of the
AndAny, which would be the match of { in the illustrated example.  This
does not work if the matches are contradictory.  In that case, the
conjunction of the contradiction leads to nothing.  This problem is
illustrated by snprintf.{cocci,c,res}, which is based on an example from
Denis Efremov.  In that example, one branch of the disjunction matches
if the identifier metavariable buf is bound to buf and the other branch
matches if this metavariable is not bound to buf.  Taking the
conjunction of these two pieces of information results in nothing.

A better solution is to more closely follow the standard encoding of a
<... P ...>, which places P v not P in the left argument of an AU.  The
only difference is that we don't care about the actual paths
(predecessors in the graph).  Instead, we can just take the reachable
nodes one by one, as if they were connected in some list whose order we
don't care about.

The changes in asttoctl2.ml address the P v not P issue.

The changes in ctl_engine.ml simulate the AU (which can also be thought
of as an AG - we want that P v not P is true at all of the nodes).

This makes the snprintf test work.

It also makes the test failing_andany work, which did not work
previously.  That test case was created the previous time this problem
was encountered.

It also makes clear that the test case incompatible_value had an
incorrect result.  The correct result can be seen by disabling the
andany optimization, ie by commenting out the line using optim1 in
asttoctl2.ml.  The original idea was that the first rule should not
match, because it would require E to have incompatible values.  But this
is not correct.  The rule can match in three ways.  One with E bound to
1, one with E bound to 2, and one with no binding of E at all.  The
first two each cause a change to be performed by the second rule.  The
third does not, because a match with the second rule requires a binding
of E and there is none.

These changes may reduce the performance benefit of the andany
optimization.  Now, as indicated by AG and P v not P, every reachable
node will have some kind of match, while previously only the nodes that
matched P had a match.  However, since we still consider every node only
once, rather than considering its connections to other nodes, and
especially since there is no risk of AW due to the presence of loops,
the optimization should still give some performance benefit.

There are a few other cases where AndAny is introduced.  One of them is
related to code added at the end of a function.  This is illustrated by
the test extra.cocci.  In that case, the code generated in asttoctl2.ml
was already correct, and the new CTL interpretation of AndAny described
above gives the right answer.  Other parts may need further attention,
as examples arise.