add newline after unminus of danger ;
When a declaration appears to be fully removed in the semantic patch, it
might not actually be removed, if it declares some extra non-matching
variables.  In that case, the declaration is kept and if code was added
to replace the declaration, it will need a newline in front of it,
analogous to the effect of giving the before argument as After to
Unparse_cocci.