recognize Comma as a fake token to be dropped by sgrep
Fake commas are kept around in the unparsing process.  Remove them
in the sgrep (*) case with the other fake tokens.

It is not possible to have a proper test case for this issue because *
doesn't change the code.  The problem can be seen using the included
test case and the command line argument --save-tmp-files on a version of
spatch from before this commit.  Looking at the generated file in /tmp
shows the extra comma.  The fake comma on this line causes a - to be
printed in the generated diff.

Reported-by: Markus Elfring <Markus.Elfring@web.de>