omit little drop functions
In the sgrep case the drop functions are not efficient because they
rebuild the affected top-level code fragment multiple times.  Some of
the functions are used in one other place.  Since we already have
functions for checking for various syntax types and since the function
name exclude is quite understandable, just inline the drop functions and
remove them.