improve pretty printing in the presence of comma expressions
Comma is a terminator in a structure or enum initialization, but not in
a comma expression (e1, e2).  In the former case, an update to the
indentation should be preserved for subsequent lines, but in the latter
case such an update should be discarded.  Add the variable instruct in
token_effect in order to distinguish between the cases.  The test
comma_call illustrates the problem.  Fixing the problem required taking
into account the tests ben and enum_assign.