avoid incrementing accumulator for calls
The accumulator is used in adjust indentation to account for the fact
that eg an if branch should be indented even if there is no surrounding
{}.  This was triggered by a ) with no following { or ; (the former
would indicate that the branch does have braces and the latter would
indicate that the ) is the end of a function call, not an if/for/while
test).  In this situation another possible token following the ) is a
comma, which also does not trigger the need for subsequent indentation.
Thus add comma to the list of tokens for which the accumulator is not
extended.