c: parsing: fix assign_expr
assign_expr did not follow the C grammar because
"in C grammar they put unary_expr, but in fact it must be
cast_expr, otherwise (int * ) xxx = &yy; is not allowed"
which is absolutely wrong. The presented construct is not
valid and the C grammar is right.

This fixes an issue discussed with Joe Perches on the mailing list

New test assign_deref_funcall_in_macro_it serves as witness