c: parsing: fix cond_expr
cond_expr did not follow the C grammar because
"in C grammar they put TDotDot cond_expr, but in fact it must be
assign_expr, otherwise pnp ? x : x = 0x388  is not allowed"
which is wrong. The presented construct is not valid and the C grammar
is right.

No new test was added because cond_expr is a subset of assign_expr in
the C grammar and, therefore, a negative test would be required to
check the exclusion of invalid constructs