Improve handleability of __attribute__ in PLUS
This patch improves the handleability of __attribute__ in PLUS.
The previous Coccinelle regarded __attribute__ in PLUS as a
comment and the comment was put just after an identifier
without a space.

Example:
$ cat /tmp/transformation.cocci
@@
identifier p;
type T;
@@
- T p __attribute__((myattr));
+ T p __attribute__((section(".shared")));

$ cat /tmp/file2.c
char hello[30] __attribute__((myattr));

$ spatch -sp_file /tmp/transformation.cocci /tmp/file2.c
init_defs_builtins: /root/hle/pub/coccinelle/standard.h
HANDLING: /tmp/file2.c
diff =
 --- /tmp/file2.c
 +++ /tmp/cocci-output-1812-cbec51-file2.c
 @@ -1 +1 @@
 -char hello[30] __attribute__((myattr));
 +char hello__attribute__((section(".shared")))[30];

 ---
This modification basically enables Coccinelle to handle the arrays with
__attribute__ in PLUS by regarding the arguments of __attribute__
as an identifier (e.g. section(".shared") is an identifier).

However, since the parser for SmPL does not support the infix-style
__attribute__ like:
char __attribute__ hello;

this patch treats these style occurences of __attribute__ specially as
a comment, which is almost the same way as previously.

Signed-off-by: Keisuke Nishimura <mumumu0722@gmail.com>