don't regenerate script value
The checking for variables used subsequently recalls the function used to
get the fresh variable value on the inherited environment.  This is
undesirable if the script code has side effects.  This change memoizes the
result of the script code.  It is not clear that this is the correct approach
if the variable is bound in many places, but it seems to be what is done
in the basic fresh id case that prompts the user.  Overall the result
in the script case seems to be no better or worse than the result in the
prompt case.

This means that no matter how many environments are considered, there will
only be one value for the fresh variable.  It is not clear whether this
is what is wanted, but it is what is done in the prompt case.

This patch also makes some whitespace cleanups, needed by adding the extra
indentation.