support operator redefinition
The problem is that operator and eg + cannot be treated by the lexer as an identifier,
because C doesn't want that.  The lexer returns one token at a time, and we would
want two tokens for C and one for C++.  So instead, we lex as operator and +,
and then inparsing hacks merge the two identifiers for the C++ case.  The merging is
done at the top of the main parsing function where there are some other
manipulations of tokens.