Fix 263: wrong default path for COCCINELLE_HOME
COCCINELLE_HOME is the directory where standard.iso is looked for.
If COCCINELLE_HOME is not defined, we consider the directory $bindir
where the current executable is. If $bindir/standard.iso exists,
we use COCCINELLE_HOME=$bindir (this is a usual case during
development, where we run spatch.opt from the working directory of the
repository).

Otherwise, we suppose that coccinelle has been installed (make
install), and that standard.iso is installed in $libdir, where
$libdir is $exec_prefix/lib.

Before this commit, we considered wrongly that $exec_prefix was equal
to $bindir, whereas the default value for $bindir is
$exec_prefix/bin. Therefore, we should take for $exec_prefix the
parent directory of $bindir.