lower menhir optimization level
Starting from menhir.20211230, menhir has a new code generation
back-end that generates by default a .ml file that is too big for
OCaml to compile. The code size is reduced is the optimization level
is reduced to -O 1 (by default, -O 2 is used).

This commit introduces a variable MENHIR_FLAGS to the Makefile, and
set MENHIR_FLAGS to -O 1 by default if the current version of menhir
supports this option (introduced in menhir.20211230, for instance the
bundled menhir version does not support it).